refactor to common-file and daft-file
