+++ Cargo.toml
@@ -1,6 +1,7 @@
 [dependencies]
 common-daft-config = {path = "src/common/daft-config", default-features = false}
 common-display = {path = "src/common/display", default-features = false}
+common-file = {path = "src/common/file", default-features = false}
 common-file-formats = {path = "src/common/file-formats", default-features = false}
 common-hashable-float-wrapper = {path = "src/common/hashable-float-wrapper", default-features = false}
 common-metrics = {path = "src/common/metrics", default-features = false}
@@ -22,6 +23,7 @@ daft-csv = {path = "src/daft-csv", default-features = false}
 daft-dashboard = {path = "src/daft-dashboard", default-features = false}
 daft-distributed = {path = "src/daft-distributed", default-features = false}
 daft-dsl = {path = "src/daft-dsl", default-features = false}
+daft-file = {path = "src/daft-file", default-features = false}
 daft-functions = {path = "src/daft-functions"}
 daft-functions-binary = {path = "src/daft-functions-binary", default-features = false}
 daft-functions-json = {path = "src/daft-functions-json", default-features = false}
@@ -69,6 +71,7 @@ python = [
   "common-display/python",
   "common-partitioning/python",
   "common-resource-request/python",
+  "common-file/python",
   "common-file-formats/python",
   "common-scan-info/python",
   "common-system-info/python",
@@ -81,6 +84,7 @@ python = [
   "daft-dashboard/python",
   "daft-distributed/python",
   "daft-dsl/python",
+  "daft-file/python",
   "daft-functions-binary/python",
   "daft-functions-json/python",
   "daft-functions-list/python",
@@ -178,6 +182,7 @@ members = [
   "src/daft-csv",
   "src/daft-dashboard",
   "src/daft-dsl",
+  "src/daft-file",
   "src/daft-functions",
   "src/daft-functions-binary",
   "src/daft-functions-json",
@@ -232,6 +237,7 @@ comfy-table = "7.1.1"
 common-daft-config = {path = "src/common/daft-config"}
 common-display = {path = "src/common/display", default-features = false}
 common-error = {path = "src/common/error", default-features = false}
+common-file = {path = "src/common/file-formats"}
 common-file-formats = {path = "src/common/file-formats"}
 common-metrics = {path = "src/common/metrics"}
 common-runtime = {path = "src/common/runtime", default-features = false}
@@ -239,6 +245,7 @@ daft-algebra = {path = "src/daft-algebra"}
 daft-context = {path = "src/daft-context"}
 daft-core = {path = "src/daft-core"}
 daft-dsl = {path = "src/daft-dsl"}
+daft-file = {path = "src/daft-file"}
 daft-functions = {path = "src/daft-functions"}
 daft-functions-binary = {path = "src/daft-functions-binary"}
 daft-functions-json = {path = "src/daft-functions-json"}
