+++ daft/daft/__init__.pyi
@@ -2133,6 +2133,10 @@ class PyDaftFile:
     def __init__(self, path: str | None = None, data: bytes | None = None) -> None: ...
     def read(self, size: int | None = None) -> bytes: ...
     def seek(self, offset: int, whence: int = 0) -> int: ...
+    @staticmethod
+    def _from_path(path: str) -> PyDaftFile: ...
+    @staticmethod
+    def _from_bytes(data: bytes) -> PyDaftFile: ...
     def tell(self) -> int: ...
     def close(self) -> None: ...
     def __enter__(self) -> PyDaftFile: ...
