+++ src/common/file/Cargo.toml
@@ -1,0 +1,22 @@
+[dependencies]
+common-error = {path = "../error", default-features = false}
+common-py-serde = {path = "../py-serde", default-features = false}
+daft-schema = {path = "../../daft-schema", default-features = false}
+pyo3 = {workspace = true, optional = true}
+serde = {workspace = true}
+
+[features]
+python = [
+  "dep:pyo3",
+  "common-error/python",
+  "common-py-serde/python",
+  "daft-schema/python"
+]
+
+[lints]
+workspace = true
+
+[package]
+edition = {workspace = true}
+name = "common-file"
+version = {workspace = true}
