+++ src/daft-core/src/lib.rs
@@ -25,7 +25,6 @@ pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
     parent.add_class::<join::JoinType>()?;
     parent.add_class::<join::JoinStrategy>()?;
     parent.add_class::<join::JoinSide>()?;
-    parent.add_class::<lit::python::PyDaftFile>()?;
 
     Ok(())
 }
