+++ src/daft-core/src/lit/mod.rs
@@ -9,6 +9,7 @@ use std::{
 };
 
 use common_error::{ensure, DaftError, DaftResult};
+use common_file::DaftFile;
 use common_hashable_float_wrapper::FloatWrapper;
 #[cfg(feature = "python")]
 use common_py_serde::PyObjectWrapper;
@@ -89,53 +90,6 @@ pub enum Literal {
     /// TODO chore: audit struct literal vs. struct expression support.
     Struct(IndexMap<Field, Literal>),
     File(DaftFile),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum DaftFile {
-    Reference(String),
-    Data(Vec<u8>),
-}
-
-impl DaftFile {
-    pub fn get_type(&self) -> DaftFileType {
-        match self {
-            Self::Reference(_) => DaftFileType::Reference,
-            Self::Data(_) => DaftFileType::Data,
-        }
-    }
-}
-
-#[repr(u8)]
-pub enum DaftFileType {
-    Reference = 0,
-    Data = 1,
-}
-
-impl TryFrom<u8> for DaftFileType {
-    type Error = DaftError;
-
-    fn try_from(value: u8) -> std::result::Result<Self, Self::Error> {
-        match value {
-            0 => Ok(Self::Reference),
-            1 => Ok(Self::Data),
-            _ => Err(DaftError::ValueError(format!(
-                "Invalid DaftFileType value: {}",
-                value
-            ))),
-        }
-    }
-}
-
-impl Eq for DaftFile {}
-
-impl Hash for DaftFile {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        match self {
-            Self::Reference(path) => path.hash(state),
-            Self::Data(bytes) => bytes.hash(state),
-        }
-    }
 }
 
 impl Eq for Literal {}
