+++ src/daft-file/Cargo.toml
@@ -1,0 +1,22 @@
+[dependencies]
+common-error = {path = "../common/error", default-features = false}
+common-file = {path = "../common/file", default-features = false}
+common-runtime = {path = "../common/runtime", default-features = false}
+daft-core = {path = "../daft-core", default-features = false}
+daft-io = {path = "../daft-io", default-features = false}
+pyo3 = {workspace = true, optional = true}
+
+[features]
+python = [
+  "common-error/python",
+  "daft-io/python",
+  "dep:pyo3"
+]
+
+[lints]
+workspace = true
+
+[package]
+name = "daft-file"
+edition.workspace = true
+version.workspace = true
